Move code into seperate files and refactor a bit
