+++ backend/src/models.py
@@ -1,7 +1,5 @@
 from sqlalchemy.orm import declarative_base, relationship
 from sqlalchemy import Column, Integer, String, Float, Date, ForeignKey, Boolean
-from flask_sqlalchemy import SQLAlchemy
-db = SQLAlchemy()
 Base = declarative_base()
 
 class Chemical(Base):
@@ -28,6 +26,11 @@ class Chemical(Base):
         foreign_keys=[Minimum_Unit_ID]
     )
     Chemical_Manufacturers = relationship("Chemical_Manufacturer", back_populates="Chemical")
+
+    def __hash__(self):
+        return hash(self.Chemical_ID)
+    def __eq__(self, other):
+        return self.Chemical_ID == other.Chemical_ID
 
 class Chemical_Manufacturer(Base):
     __tablename__ = 'Chemical_Manufacturer'
